Separate routing detection from App into its own middleware
